Lay groundwork for reading source code files.
This is the basic mechanism for which source code will be read into the
compiler. A Source object will be able to pass through the compiler and
be turned into a translation unit.

Source objects should also be producible from the command line as
strings, or in a REPL, not just from files.