Start implementing Lexical Analysis.
This is the beginnings of the lexical analysis for the compiler. The
lexer should consume the source and produce a token stream at the end of
it.

Currently this has only had comment and string consumers added to it.

In addition a test file has been added "string.bolt" which allows
testing the lexer in its current state.