Generate _Hello World_ IR code!
This is the first time the compiler has generated a _Hello World_ IR
code program. It marks a major milestone, but still not v0.0.1 as it is
not reliable.

There needs to be additions to semantic analysis. Void returns should be
automatically synthesised by the compiler, and errors for missing
returns should be given. Currently to work around the return issues I've
added an explicit return into the `print()` function of the standard
library, which isn't really a pleasant thing for users to have to do.