Allow lookup of expression locations.
AST Expressions need to be able to expose common values such as location
without having to unwrap/check/compare each one every time. This was
initially the reason for opting for the solution I did.

This adds a property to the enum which provides the location in a
convenient manner.