Start adding function code generation.
This is the beginning of code generation for functions. It is currently
only capable to producing IR code that declares a function. Function
definitions still need to be produced. However this will require other
code generators to be put in place to be able to produce the body of a
function.