Generate Constants in IR Output.
This has involved a number of additions to the code generator, including
the ability to synthesize "label" names for anonymous/unnamed values in
the input source code.

Additionally this has introduced the concept of Global value emitters
and emitters for string and integer literals. These are still relatively
basic mirroring what was previously present in the old AST
implementation.